requirements.txt
numba==0.49.1
librosa==0.7.2
torch>=1.2.0
phonemizer>=2.2
praat-parselmouth==0.3.3
parsel
scipy
tensorboard
matplotlib
unidecode
inflect
